Create script for Sending Email
This is the code for receiving emails via Arduino WiFi connection. In main_code, the url and recipient are silently replaced with "----". This is for personal information security. To use the code, distribute the project from the following site: https://developers.google.com/apps-script?hl=ko, insert the url, and insert the email address to receive in the recipient.